[Refactor] Move run_app & build_app to cli
