+++ pyfyre_cli/run_app.py
@@ -8,7 +8,7 @@
 
 import os
 from livereload import Server
-from build import build_app, create_pages, bundle_scripts, add_cpython_packages
+from pyfyre_cli.build_app import build_app, create_pages, bundle_scripts, add_cpython_packages
 
 
 def _build_src() -> None:
@@ -20,13 +20,9 @@ def _build_settings() -> None:
 	bundle_scripts(production=False)
 	add_cpython_packages()
 
-
-if __name__ == "__main__":
-	if os.path.dirname(__file__) == os.getcwd():
-		build_app()
-		server = Server()
-		server.watch("src/", _build_src)
-		server.watch("settings.py", _build_settings)
-		server.serve(root="public")
-	else:
-		print("You must be in the directory of the project to run it.")
+def run_app() -> None:
+    build_app()
+    server = Server()
+    server.watch(os.path.join(os.getcwd(), 'src'), _build_src)
+    server.watch(os.path.join(os.getcwd(), 'settings.py'), _build_settings)
+    server.serve(root=os.path.join(os.getcwd(), 'public'))
