+++ pyfyre_cli/user_templates/default/public/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<title>{title}</title>
+		
+		<meta charset="utf-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=yes" />
+		
+		<link rel="icon" href="{icon}" />
+		<link rel="stylesheet" href="/style.css" />
+		
+		<!-- Start of Brython -->
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/brython/3.10.7/brython.min.js"></script>
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/brython/3.10.7/brython_stdlib.min.js"></script>
+		<script src="/_pyfyre/src.brython.js"></script>
+		<script type="text/python">
+			import pyfyre
+			pyfyre.PRODUCTION = {prod_env}
+			import index
+		</script>
+		<!-- End of Brython -->
+		
+		{head}
+	</head>
+	<body onload="brython()"></body>
+</html>
