+++ pyfyre_cli/user_templates/default/public/style.css
